% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.cart.R
\name{mice.pcr.sim.impute.cart}
\alias{mice.pcr.sim.impute.cart}
\alias{cart}
\title{Imputation by classification and regression trees}
\usage{
mice.pcr.sim.impute.cart(y, ry, x, wy = NULL, minbucket = 5, cp = 1e-04, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{minbucket}{The minimum number of observations in any terminal node used.
See \code{\link{rpart.control}} for details.}

\item{cp}{Complexity parameter. Any split that does not decrease the overall
lack of fit by a factor of cp is not attempted. See \code{\link{rpart.control}}
for details.}

\item{...}{Other named arguments passed down to \code{rpart()}.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}

Numeric vector of length \code{sum(!ry)} with imputations
}
\description{
Imputes univariate missing data using classification and regression trees.
}
\details{
Imputation of \code{y} by classification and regression trees. The procedure
is as follows:
\enumerate{
\item Fit a classification or regression tree by recursive partitioning;
\item For each \code{ymis}, find the terminal node they end up according to the fitted tree;
\item Make a random draw among the member in the node, and take the observed value from that
draw as the imputation.
}
}
\examples{
require(rpart)

imp <- mice.pcr.sim(nhanes2, meth = "cart", minbucket = 4)
plot(imp)
}
\references{
Doove, L.L., van Buuren, S., Dusseldorp, E. (2014), Recursive partitioning
for missing data imputation in the presence of interaction Effects.
Computational Statistics \& Data Analysis, 72, 92-104.

Breiman, L., Friedman, J. H., Olshen, R. A., and Stone, C. J.
(1984), Classification and regression trees, Monterey, CA: Wadsworth &
Brooks/Cole Advanced Books & Software.

Van Buuren, S. (2018).
\href{https://stefvanbuuren.name/fimd/sec-cart.html}{\emph{Flexible Imputation of Missing Data. Second Edition.}}
Chapman & Hall/CRC. Boca Raton, FL.
}
\seealso{
\code{\link{mice.pcr.sim}}, \code{\link{mice.pcr.sim.impute.rf}},
\code{\link[rpart]{rpart}}, \code{\link[rpart]{rpart.control}}

Other univariate imputation functions: 
\code{\link{mice.pcr.sim.impute.durr.bern}()},
\code{\link{mice.pcr.sim.impute.durr.gaus}()},
\code{\link{mice.pcr.sim.impute.iurr.gaus}()},
\code{\link{mice.pcr.sim.impute.lda}()},
\code{\link{mice.pcr.sim.impute.logreg.boot}()},
\code{\link{mice.pcr.sim.impute.logreg}()},
\code{\link{mice.pcr.sim.impute.mean}()},
\code{\link{mice.pcr.sim.impute.midastouch}()},
\code{\link{mice.pcr.sim.impute.mnar.logreg}()},
\code{\link{mice.pcr.sim.impute.norm.boot}()},
\code{\link{mice.pcr.sim.impute.norm.nob}()},
\code{\link{mice.pcr.sim.impute.norm.predict}()},
\code{\link{mice.pcr.sim.impute.norm}()},
\code{\link{mice.pcr.sim.impute.pcr.boot}()},
\code{\link{mice.pcr.sim.impute.pcr.mixed}()},
\code{\link{mice.pcr.sim.impute.pmm}()},
\code{\link{mice.pcr.sim.impute.polr}()},
\code{\link{mice.pcr.sim.impute.polyreg}()},
\code{\link{mice.pcr.sim.impute.quadratic}()},
\code{\link{mice.pcr.sim.impute.rf}()},
\code{\link{mice.pcr.sim.impute.ri}()}
}
\author{
Lisa Doove, Stef van Buuren, Elise Dusseldorp, 2012
}
\concept{univariate imputation functions}
\keyword{datagen}
