% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mice.impute.pcr.mixed.R
\name{mice.pcr.sim.impute.pcr.mixed}
\alias{mice.pcr.sim.impute.pcr.mixed}
\alias{pcr.mixed}
\title{Imputation by principal component regression}
\usage{
mice.pcr.sim.impute.pcr.mixed(y, ry, x, wy = NULL, npcs, ...)
}
\arguments{
\item{y}{Vector to be imputed}

\item{ry}{Logical vector of length \code{length(y)} indicating the
the subset \code{y[ry]} of elements in \code{y} to which the imputation
model is fitted. The \code{ry} generally distinguishes the observed
(\code{TRUE}) and missing values (\code{FALSE}) in \code{y}.}

\item{x}{Numeric design matrix with \code{length(y)} rows with predictors for
\code{y}. Matrix \code{x} may have no missing values.}

\item{wy}{Logical vector of length \code{length(y)}. A \code{TRUE} value
indicates locations in \code{y} for which imputations are created.}

\item{npcs}{The number of principal components to extract with the pcr.mixed method.}

\item{...}{Other named arguments.}
}
\value{
Vector with imputed data, same type as \code{y}, and of length
\code{sum(wy)}
}
\description{
Imputes univariate missing data by extracting PCs from all auxiliary variables.
This version uses a correlation matrix estiamted with the mixedCor()
for a mixture of tetrachorics, polychorics, Pearsons, biserials, and polyserials,
as appropriate.
}
\details{
Extracts components from x and uses them as predictors the regular
norm.boot() fashion.
}
\references{
Jolliffe, I. (2002) Principal Component Analysis (2nd ed). Springer.
}
\seealso{
Other univariate imputation functions: 
\code{\link{mice.pcr.sim.impute.cart}()},
\code{\link{mice.pcr.sim.impute.durr.bern}()},
\code{\link{mice.pcr.sim.impute.durr.gaus}()},
\code{\link{mice.pcr.sim.impute.iurr.gaus}()},
\code{\link{mice.pcr.sim.impute.lda}()},
\code{\link{mice.pcr.sim.impute.logreg.boot}()},
\code{\link{mice.pcr.sim.impute.logreg}()},
\code{\link{mice.pcr.sim.impute.mean}()},
\code{\link{mice.pcr.sim.impute.midastouch}()},
\code{\link{mice.pcr.sim.impute.mnar.logreg}()},
\code{\link{mice.pcr.sim.impute.norm.boot}()},
\code{\link{mice.pcr.sim.impute.norm.nob}()},
\code{\link{mice.pcr.sim.impute.norm.predict}()},
\code{\link{mice.pcr.sim.impute.norm}()},
\code{\link{mice.pcr.sim.impute.pcr.boot}()},
\code{\link{mice.pcr.sim.impute.pcr.pmm}()},
\code{\link{mice.pcr.sim.impute.pmm}()},
\code{\link{mice.pcr.sim.impute.polr}()},
\code{\link{mice.pcr.sim.impute.polyreg}()},
\code{\link{mice.pcr.sim.impute.quadratic}()},
\code{\link{mice.pcr.sim.impute.rf}()},
\code{\link{mice.pcr.sim.impute.ri}()}
}
\author{
Edoardo Costantini, 2021
}
\concept{univariate imputation functions}
\keyword{datagen}
